Add props for UpcomingTemplate, move the mock data to +page
